+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     i.e `b := { sum(x) >= 10 || x in CS.AllDifferentSet() }`
 - Support for `&&` and `||` outside of `Indicator` and `Reified`
   - i.e `sum(x) >= 10 || x in CS.AllDifferentSet()`
+- Basic support for `Indicator` inside of reified:
+  - i.e `@constraint(model, b1 := {b2 => { v == 1 }})`
+  - currently lacks support for bridges such that `v > 1` in the inner constraint will fail
 - Some general performance updates [PR #247](https://github.com/Wikunia/ConstraintSolver.jl/pull/247)
 
 ## v0.6.3 (17th of January 2021)